requirements.txt
# ============================================
# RUNZA BACKEND - DEPENDENCIAS
# ============================================

# FastAPI y servidor
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Base de datos
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Validación y serialización
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0.post1

# Autenticación
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Utilidades
python-dotenv==1.0.1
httpx==0.26.0

# Testing (versiones compatibles)
pytest==7.4.4
pytest-asyncio==0.23.4
pytest-cov==4.1.0

# Desarrollo
black==24.1.1
isort==5.13.2
flake8==7.0.0

mediapipe==0.10.9
opencv-python-headless==4.8.1.78
numpy>=1.24.0
Pillow>=10.0.0